# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CareerPilot Interview API",
    "version": "0.0.0-dev",
    "description": "REST API for CareerPilot Interview by CodeBegun. Versioned endpoints live under /api/v1. Successful responses are `{ data }`; errors are `{ error: { code, message, details?, requestId } }`."
  },
  "servers": [
    {
      "url": "https://api.interview.codebegun.com"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      },
      "refreshCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "cbi_rt",
        "description": "httpOnly refresh cookie (`cbi_rt` / `cbi_admin_rt`; `__Secure-` prefix over HTTPS). Requests must also send `X-CB-CSRF: 1`."
      }
    },
...
  },
  "paths": {
...
    "/api/v1/auth/otp/request": {
      "post": {
        "tags": [
          "Candidate auth"
        ],
        "summary": "Send a one-time code by email or SMS",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "channel": {
                    "type": "string",
                    "enum": [
                      "EMAIL",
                      "MOBILE"
                    ]
                  },
                  "destination": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 254
                  }
                },
                "required": [
                  "channel",
                  "destination"
                ]
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "challengeId": {
                          "type": "string"
                        },
                        "sentTo": {
                          "type": "string"
                        },
                        "expiresAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "resendAvailableAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      },
                      "required": [
                        "challengeId",
                        "sentTo",
                        "expiresAt",
                        "resendAvailableAt"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "BAD_REQUEST",
                            "VALIDATION_FAILED",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "ORIGIN_NOT_ALLOWED",
                            "SERVICE_UNAVAILABLE",
                            "INTERNAL_ERROR",
                            "CSRF_REJECTED",
                            "OTP_INVALID",
                            "OTP_EXPIRED",
                            "OTP_TOO_MANY_ATTEMPTS",
                            "OTP_COOLDOWN",
                            "IDENTITY_IN_USE",
                            "ACCOUNT_SUSPENDED",
                            "FEATURE_DISABLED",
                            "PROVIDER_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "BAD_REQUEST",
                            "VALIDATION_FAILED",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "ORIGIN_NOT_ALLOWED",
                            "SERVICE_UNAVAILABLE",
                            "INTERNAL_ERROR",
                            "CSRF_REJECTED",
                            "OTP_INVALID",
                            "OTP_EXPIRED",
                            "OTP_TOO_MANY_ATTEMPTS",
                            "OTP_COOLDOWN",
                            "IDENTITY_IN_USE",
                            "ACCOUNT_SUSPENDED",
                            "FEATURE_DISABLED",
                            "PROVIDER_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "503": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "BAD_REQUEST",
                            "VALIDATION_FAILED",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "ORIGIN_NOT_ALLOWED",
                            "SERVICE_UNAVAILABLE",
                            "INTERNAL_ERROR",
                            "CSRF_REJECTED",
                            "OTP_INVALID",
                            "OTP_EXPIRED",
                            "OTP_TOO_MANY_ATTEMPTS",
                            "OTP_COOLDOWN",
                            "IDENTITY_IN_USE",
                            "ACCOUNT_SUSPENDED",
                            "FEATURE_DISABLED",
                            "PROVIDER_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/otp/verify": {
      "post": {
        "tags": [
          "Candidate auth"
        ],
        "summary": "Verify the code and start a session (sets the refresh cookie)",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "challengeId": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 64
                  },
                  "code": {
                    "type": "string",
                    "pattern": "^\\d{6}$"
                  }
                },
                "required": [
                  "challengeId",
                  "code"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "accessToken": {
                          "type": "string"
                        },
                        "accessTokenExpiresAt": {
                          "type": "string",
                          "format": "date-time"
                        },
                        "user": {
                          "type": "object",
                          "properties": {
                            "id": {
                              "type": "string"
                            },
                            "email": {
                              "type": [
                                "string",
                                "null"
                              ],
                              "format": "email"
                            },
                            "mobile": {
                              "type": [
                                "string",
                                "null"
                              ]
                            },
                            "status": {
                              "type": "string",
                              "enum": [
                                "ACTIVE",
                                "SUSPENDED"
                              ]
                            },
                            "adminRoles": {
                              "type": "array",
                              "items": {
                                "type": "string",
                                "enum": [
                                  "SUPER_ADMIN",
                                  "OPERATIONS_ADMIN",
                                  "CONTENT_ADMIN",
                                  "SUPPORT_ADMIN",
                                  "FINANCE_ADMIN"
                                ]
                              }
                            },
                            "onboardingCompleted": {
                              "type": "boolean"
                            },
                            "profile": {
                              "type": "object",
                              "properties": {
                                "displayName": {
                                  "type": [
                                    "string",
                                    "null"
                                  ]
                                },
                                "preferredInterviewLanguage": {
                                  "type": "string",
                                  "enum": [
                                    "auto",
                                    "en",
                                    "hi",
                                    "te"
                                  ]
                                },
                                "experienceLevel": {
                                  "type": [
                                    "string",
                                    "null"
                                  ],
                                  "enum": [
                                    "STUDENT",
                                    "FRESHER",
                                    "EARLY_CAREER",
                                    "MID_LEVEL",
                                    "SENIOR",
                                    null
                                  ]
                                },
                                "currentRole": {
                                  "type": [
                                    "string",
                                    "null"
                                  ]
                                },
                                "productUpdatesOptIn": {
                                  "type": "boolean"
                                }
                              },
                              "required": [
                                "displayName",
                                "preferredInterviewLanguage",
                                "experienceLevel",
                                "currentRole",
                                "productUpdatesOptIn"
                              ]
                            },
                            "identities": {
                              "type": "array",
                              "items": {
                                "type": "object",
                                "properties": {
                                  "provider": {
                                    "type": "string",
                                    "enum": [
                                      "EMAIL",
                                      "MOBILE",
                                      "GOOGLE"
                                    ]
                                  },
                                  "display": {
                                    "type": "string"
                                  },
                                  "verifiedAt": {
                                    "type": "string",
                                    "format": "date-time"
                                  }
                                },
                                "required": [
                                  "provider",
                                  "display",
                                  "verifiedAt"
                                ]
                              }
                            },
                            "createdAt": {
                              "type": "string",
                              "format": "date-time"
                            }
                          },
                          "required": [
                            "id",
                            "email",
                            "mobile",
                            "status",
                            "adminRoles",
                            "onboardingCompleted",
                            "profile",
                            "identities",
                            "createdAt"
                          ]
                        }
                      },
                      "required": [
                        "accessToken",
                        "accessTokenExpiresAt",
                        "user"
                      ]
                    }
                  },
                  "required": [
                    "data"
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "BAD_REQUEST",
                            "VALIDATION_FAILED",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "ORIGIN_NOT_ALLOWED",
                            "SERVICE_UNAVAILABLE",
                            "INTERNAL_ERROR",
                            "CSRF_REJECTED",
                            "OTP_INVALID",
                            "OTP_EXPIRED",
                            "OTP_TOO_MANY_ATTEMPTS",
                            "OTP_COOLDOWN",
                            "IDENTITY_IN_USE",
                            "ACCOUNT_SUSPENDED",
                            "FEATURE_DISABLED",
                            "PROVIDER_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "BAD_REQUEST",
                            "VALIDATION_FAILED",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "ORIGIN_NOT_ALLOWED",
                            "SERVICE_UNAVAILABLE",
                            "INTERNAL_ERROR",
                            "CSRF_REJECTED",
                            "OTP_INVALID",
                            "OTP_EXPIRED",
                            "OTP_TOO_MANY_ATTEMPTS",
                            "OTP_COOLDOWN",
                            "IDENTITY_IN_USE",
                            "ACCOUNT_SUSPENDED",
                            "FEATURE_DISABLED",
                            "PROVIDER_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          },
          "429": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "BAD_REQUEST",
                            "VALIDATION_FAILED",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "ORIGIN_NOT_ALLOWED",
                            "SERVICE_UNAVAILABLE",
                            "INTERNAL_ERROR",
                            "CSRF_REJECTED",
                            "OTP_INVALID",
                            "OTP_EXPIRED",
                            "OTP_TOO_MANY_ATTEMPTS",
                            "OTP_COOLDOWN",
                            "IDENTITY_IN_USE",
                            "ACCOUNT_SUSPENDED",
                            "FEATURE_DISABLED",
                            "PROVIDER_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/auth/logout": {
      "post": {
        "tags": [
          "Candidate auth"
        ],
        "summary": "End this session",
        "security": [
          {
            "refreshCookie": []
          }
        ],
        "responses": {
          "204": {
            "description": "No content"
          },
          "403": {
            "description": "Error",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "object",
                      "properties": {
                        "code": {
                          "type": "string",
                          "enum": [
                            "BAD_REQUEST",
                            "VALIDATION_FAILED",
                            "UNAUTHENTICATED",
                            "FORBIDDEN",
                            "NOT_FOUND",
                            "CONFLICT",
                            "PAYLOAD_TOO_LARGE",
                            "RATE_LIMITED",
                            "ORIGIN_NOT_ALLOWED",
                            "SERVICE_UNAVAILABLE",
                            "INTERNAL_ERROR",
                            "CSRF_REJECTED",
                            "OTP_INVALID",
                            "OTP_EXPIRED",
                            "OTP_TOO_MANY_ATTEMPTS",
                            "OTP_COOLDOWN",
                            "IDENTITY_IN_USE",
                            "ACCOUNT_SUSPENDED",
                            "FEATURE_DISABLED",
                            "PROVIDER_UNAVAILABLE"
                          ]
                        },
                        "message": {
                          "type": "string"
                        },
                        "details": {},
                        "requestId": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "code",
                        "message"
                      ]
                    }
                  },
                  "required": [
                    "error"
                  ]
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}